Update flash-it.sh
adding towboot switch

but there was more to search as expected
  had an error: (Unmounting ./sdcard.img; umount: ./sdcard.img: not mounted.)
  and (..pinephone/sfos/root: special device /dev/loopp2 does not exist.)
this was causing it
  fixing sdcard size to 8GB: generating a 4GB sdcard.img and trying to make a ~6GB rootfs does not work an was te reason for the errors

in 'raw' mode: some umount errors at the end of script fixed with umount '-q' option - maybe dirty. 
  DEVICE_NODE var is used for three different cases ('raw', './sdcard.img' and '/dev/loop$LOOP_NODE' - thats many code reading on changes) and that is why it says at the end: 'Flashing /dev/loop0 OK!'

question: encryption devices.ini in the root archive from pinephone (v1.2) is '/dev/mmcblk2p3', that will not work if ist is flashed to emmc. I think the script should set i on an answer with 'n' too, but vice versa to the answer with 'y'?
  
This is still not optimal:
- after dd'ing the sdcard.img to a real sdcard the user should resize home partition before starting sfos. Or the home partition should be resized on first start.
and
Creating EXT4 file system...
Warning: The resulting partition is not properly aligned for best performance: 62500s % 2048s != 0s
Warning: The resulting partition is not properly aligned for best performance: 500001s % 2048s != 0s
Warning: The resulting partition is not properly aligned for best performance: 12207032s % 2048s != 0s